Intermediate/Hurst_Dots/requirements.txt
autopep8==1.6.0
colorgram.py==1.2.0
Pillow==9.1.1
pycodestyle==2.8.0
toml==0.10.2
